+++ crates/sui-json-rpc/src/api/governance.rs
@@ -7,13 +7,11 @@ use jsonrpsee_proc_macros::rpc;
 use sui_json_rpc_types::{BigInt, DelegatedStake, SuiCommittee};
 use sui_open_rpc_macros::open_rpc;
 use sui_types::base_types::{ObjectID, SuiAddress};
-
 use sui_types::committee::EpochId;
-
 use sui_types::sui_system_state::sui_system_state_summary::SuiSystemStateSummary;
 
-#[open_rpc(namespace = "sui", tag = "Governance Read API")]
-#[rpc(server, client, namespace = "sui")]
+#[open_rpc(namespace = "suix", tag = "Governance Read API")]
+#[rpc(server, client, namespace = "suix")]
 pub trait GovernanceReadApi {
     /// Return one or more [DelegatedStake]. If a Stake was withdrawn its status will be Unstaked.
     #[method(name = "getStakesByIds")]
