+++ crates/sui-json-rpc/src/api/mod.rs
@@ -1,42 +1,41 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-mod coin;
-mod event;
-mod extended;
-mod governance;
-mod read;
-mod transaction_builder;
-mod write;
-
 use anyhow::anyhow;
+
 pub use coin::CoinReadApiClient;
 pub use coin::CoinReadApiOpenRpc;
 pub use coin::CoinReadApiServer;
-
 pub use extended::ExtendedApiClient;
 pub use extended::ExtendedApiOpenRpc;
 pub use extended::ExtendedApiServer;
-
-pub use event::EventReadApiClient;
-pub use event::EventReadApiOpenRpc;
-pub use event::EventReadApiServer;
-
+pub use governance::GovernanceReadApiClient;
+pub use governance::GovernanceReadApiOpenRpc;
+pub use governance::GovernanceReadApiServer;
+pub use indexer::IndexerApiClient;
+pub use indexer::IndexerApiOpenRpc;
+pub use indexer::IndexerApiServer;
+pub use move_utils::MoveUtilsClient;
+pub use move_utils::MoveUtilsOpenRpc;
+pub use move_utils::MoveUtilsServer;
+pub use read::ReadApiClient;
+pub use read::ReadApiOpenRpc;
+pub use read::ReadApiServer;
+pub use transaction_builder::TransactionBuilderClient;
+pub use transaction_builder::TransactionBuilderOpenRpc;
+pub use transaction_builder::TransactionBuilderServer;
 pub use write::WriteApiClient;
 pub use write::WriteApiOpenRpc;
 pub use write::WriteApiServer;
 
-pub use governance::GovernanceReadApiClient;
-pub use governance::GovernanceReadApiOpenRpc;
-pub use governance::GovernanceReadApiServer;
-
-pub use read::ReadApiClient;
-pub use read::ReadApiOpenRpc;
-pub use read::ReadApiServer;
-
-pub use transaction_builder::TransactionBuilderClient;
-pub use transaction_builder::TransactionBuilderOpenRpc;
-pub use transaction_builder::TransactionBuilderServer;
+mod coin;
+mod extended;
+mod governance;
+mod indexer;
+mod move_utils;
+mod read;
+mod transaction_builder;
+mod write;
 
 /// Maximum number of events returned in an event query.
 /// This is equivalent to EVENT_QUERY_MAX_LIMIT in `sui-storage` crate.
