+++ crates/sui-json-rpc/src/lib.rs
@@ -28,9 +28,10 @@ pub mod api;
 mod balance_changes;
 pub mod coin_api;
 pub mod error;
-pub mod event_api;
 pub mod governance_api;
+pub mod indexer_api;
 mod metrics;
+pub mod move_utils;
 mod object_changes;
 pub mod read_api;
 mod routing_layer;
